example/requirements.txt
django>=1.4
django-form-admin
django-ajax-selects
